Added channel exceptions

+++ TestBot/src/main/java/me/bartvv/testbot/listener/CommandListener.java
@@ -1,5 +1,8 @@
 package me.bartvv.testbot.listener;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import me.bartvv.testbot.Utils;
 import me.bartvv.testbot.commands.CommandHandler;
 import net.dv8tion.jda.core.entities.Message;
@@ -11,9 +14,13 @@ import net.dv8tion.jda.core.hooks.ListenerAdapter;
 public class CommandListener extends ListenerAdapter {
 
 	private CommandHandler commandHandler;
+	private List< String > exclusions;
 
 	public CommandListener() {
 		this.commandHandler = new CommandHandler();
+		this.exclusions = new ArrayList<>();
+		this.exclusions.add( Utils.getCommand() + "close" );
+		this.exclusions.add( Utils.getCommand() + "support" );
 	}
 
 	@Override
@@ -38,6 +45,12 @@ public class CommandListener extends ListenerAdapter {
 			Utils.sendMessage( channel, "This is not a command, please only use commands in this channel" );
 			return;
 		} else {
+			String[] args = rawMessage.split( " " );
+			if ( args.length >= 1 && this.exclusions.contains( args[ 0 ].toLowerCase() ) ) {
+				this.commandHandler.processCommand( e );
+				return;
+			}
+
 			if ( rawMessage.startsWith( Utils.getCommand() ) ) {
 				message.delete().complete();
 				Utils.sendMessage( channel, "Please refrain from using other channel's then the "
